+++ IUNowPlayingAlbumBackViewController.m
@@ -10,7 +10,7 @@
 
 @implementation IUNowPlayingAlbumBackViewController (MusicGestures)
 
--(void)swipableView:(id)view pinchedToScale:(float)scale withVelocity:(float)velocity {
+-(void)swipableView:(id)view pinchedToScale:(double)scale withVelocity:(double)velocity {
 
   [self performActionForKey:kGestureBackPinch];
 
@@ -18,8 +18,8 @@
 
 -(void)swipableView:(id)view 
        pannedInDirection:(int)direction 
-       withDistance:(float)distance 
-       withVelocity:(float)velocity {
+       withDistance:(double)distance 
+       withVelocity:(double)velocity {
 
   switch(direction) {
