+++ IUNowPlayingAlbumFrontViewController.h
@@ -9,11 +9,11 @@
 @end
 
 @interface IUNowPlayingAlbumFrontViewController (MusicGestures)
--(void)swipableView:(id)view pinchedToScale:(float)scale withVelocity:(float)velocity;
+-(void)swipableView:(id)view pinchedToScale:(double)scale withVelocity:(double)velocity;
 -(void)swipableView:(id)view 
        pannedInDirection:(int)direction 
-       withDistance:(float)distance 
-       withVelocity:(float)velocity;
+       withDistance:(double)distance 
+       withVelocity:(double)velocity;
 -(void)swipableView:(id)view
        longPressedInState:(UIGestureRecognizerState)state;
 @end
